Add configurable delay before archiving
When teslausb comes in wifi range it disconnects the drives from the car,
the equivalent of physically disconnecting them. This results in possibly
truncated recordings, and music played from USB immediately stopping.
This change adds a configurable delay between connecting to wifi and starting
the archive operation.
The default delay is 20 seconds.